fix: preserve wait absent restart diagnostics
