+++ packages/maestro/src/internal/__tests__/export-label-projection.test.ts
@@ -41,12 +41,20 @@ test('keeps compound selectors that include label as hard export errors', () => 
 });
 
 test('does not export strict wait absent as Maestro notVisible', () => {
-  expect(() =>
+  let thrown: unknown;
+  try {
     exportReplayActionsToMaestro([action('wait', ['absent', 'label="Removed"', '1000'])], {
       resolveSelector: (expression) =>
         projectSelectorExpression(expression, MAESTRO_SELECTOR_PROJECTION),
-    }),
-  ).toThrow(/wait absent.*unsupported|strict.*absence|notVisible/i);
+    });
+  } catch (error) {
+    thrown = error;
+  }
+
+  expect(thrown).toBeInstanceOf(AppError);
+  if (!(thrown instanceof AppError)) return;
+  expect(thrown.message).toMatch(/unsupported|strict absence/i);
+  expect(thrown.message).not.toMatch(/notVisible/i);
 });
 
 function action(command: string, positionals: string[]): SessionAction {
