+++ src/commands/interaction/runtime/wait-stable.test.ts
