+++ src/daemon/__tests__/system-surface-disclosure.test.ts
@@ -2,7 +2,8 @@ import { test, expect, vi, beforeEach } from 'vitest';
 import { legacyDispatchCapture } from './legacy-snapshot-capture-fixture.ts';
 import { handleFindCommands } from '../interaction/index.ts';
 import { getRuntimeBindings } from './interaction-get-runtime-fixture.ts';
-import { dispatchFindReadOnlyViaRuntime, dispatchWaitViaRuntime } from '../selector-runtime.ts';
+import { dispatchFindReadOnlyViaRuntime } from '../selector-runtime.ts';
+import { dispatchWaitViaRuntime } from '../wait-runtime.ts';
 import type { DaemonRequest, DaemonResponse } from '../types.ts';
 import { ANDROID_SYSTEM_SURFACE_DISCLOSURE } from '../../core/android-system-surface-disclosure.ts';
 import { snapshotRuntimeFixture } from './snapshot-runtime-fixture.ts';
