+++ src/daemon/handlers/__tests__/wait-landmark-recording.test.ts
@@ -13,7 +13,7 @@
  */
 import { test, expect, vi, beforeEach } from 'vitest';
 import { legacyDispatchCapture } from '../../__tests__/legacy-snapshot-capture-fixture.ts';
-import { dispatchWaitViaRuntime } from '../../selector-runtime.ts';
+import { dispatchWaitViaRuntime } from '../../wait-runtime.ts';
 import type { DaemonRequest } from '../../types.ts';
 import { WAIT_LANDMARK_MISMATCH_REASON } from '@agent-device/contracts/replay';
 import type { TargetAnnotationV1 } from '@agent-device/contracts/replay';
